+++ medtog.php
@@ -1,28 +1,24 @@
 <?php
-// Toggling the lights in the meditation corner with one button.
+# Toggling the lights in the meditation corner with one button.
+# Now with multiple colour temperatures!!
 include 'functions.php';
-$lamp1 = 9; // Just on/off, please
-$lamp2 = 33; // Hue capable. Do that.
+$lamp1 = 9; # Just on/off, please
+$lamp2 = 33; # Hue capable. Do that.
 
-// First, the Shrine Lamp ($lamp1)
-if (restoreState($lamp1)) {
-    // Restore state worked. Nothing more to do here?
-}
-else {
+if(!checkState($lamp2)) { # Let's key this on the downlight
     saveOnState($lamp1);
+    saveCTState($lamp2);
     oneOn($lamp1);
+    oneOnSpotWarm($lamp2);
+    
 }
-
-// Then, the downlight ($lamp2)...
-if (restoreState($lamp2)) {
-    // Restored. Done
+elseif(checkstate($lamp2) && (getCTState($lamp2) == $ctWarm)) {
+    # $lamp1 is already on. Leave it alone.
+    oneOnSpotCool($lamp2);
 }
 else {
-    saveHueState($lamp2);
-    // Todo: make this a multi-mode toggle? Sometimes I like it warm, sometimes
-    // cool; it'd be nice to toggle these on-the-fly.
-    oneOnSpotWarm($lamp2);
-    //oneOnSpotCool($lamp2);
+    restoreState($lamp1);
+    restoreState($lamp2);
 }
 
 if(strpos($_SERVER['HTTP_USER_AGENT'], 'WebKit')) header('Location: ' . $_SERVER['medtog.php']);