+++ normal.php
@@ -14,24 +14,26 @@
 
 <?php
 include 'functions.php';
+$lightlist = array(11,12,13,14,15);
 
-oneOnBright(11);
-oneOnBright(12);
-oneOnBright(13);
-oneOnBright(14);
-oneOnBright(15);
+# Let's add a quick check whether the downlights should be full-brightness
+$full = FALSE; # We default to half-brightness
+$fulldl = $_GET['fulldl'];
+if(isset ($fulldl)) $full = TRUE; #we only care if fulldl exists; value doesn't matter
+
+foreach($lightlist as $id) {
+  oneOnBright($id);
+}
 oneOn(17);
 oneOn(20);
-//$output = `/usr/bin/env curl -s http://max.kenkl.org/lights/lrdlonfullwarm.php`;
-//$output = `/usr/bin/env curl -s http://max.kenkl.org/lights/lrdl50.php`;
 
-// Let's try something different...
+# Downlights 
 for($id = 22; $id <=25; $id++) {
     oneOn($id);
-    setCT($id, 400);
-    setLevel($id, 127);
+    setCT($id, $ctWarm);
+    if($full) setLevel($id, 254); # DLs full-brightness?
+    else setLevel($id, 127);
 }
-
 
 if(strpos($_SERVER['HTTP_USER_AGENT'], 'WebKit')) header('Location: ' . $_SERVER['normal.php']);